loss

+++ khd/kernels/swiglu/cuda_implementation/kernels.cu
@@ -140,7 +140,7 @@ __global__ void _swiglu_backward_cuda_kernel(const scalar_t *gate,
     }
 }
 
-void swiglu_forward_cuda(torch::Tensor gate, torch::Tensor up, torch::Tensor output, const int BLOCK_SIZE) {
+void swiglu_forward_cuda(torch::Tensor gate, torch::Tensor up, torch::Tensor output, const int64_t BLOCK_SIZE) {
     const int num_elements = gate.numel();
 
     AT_DISPATCH_CUSTOM_FLOAT_TYPES(
